+++ src/main/java/duke/Checker.java
@@ -1,3 +1,5 @@
+package duke;
+
 import java.time.LocalDate;
 
 public class Checker {
